deleted api/posts and fixed game.js
